Implement support for automatic re-resolution of the dest endpoint
Upon startup, periodically retries DNS until the configured host resolves successfully, dropping data in the meantime.
If the host changes in DNS, automatically switches updates to the new endpoint.
If there are multiple DNS records, a random entry will be selected whenever the list is modified.
If a DNS lookup which previously succeeded is now failing, continues sending data to the last-successful endpoint.