Fix segfault in sync_util if DNS is dead
sync_util's runner function was blindly returning a value, even if the
caller was no longer there to receive it.

This would result in a segfault when the runner function attempts to
access variables that were scoped inside the caller's stack.

To fix this, sync_util now has a 'ticketing' system which allows the
runner function to check that the caller is still waiting for the
result, before attempting to pass it back.