Only fetch agent IP once
Per DC/OS docs this shouldn't change on us:

Important: The IP address of a node must not change after DC/OS is
installed on the node. For example, the IP address must not change
when a node is rebooted or if the DHCP lease is renewed. If the IP
address of a node does change, the node must be wiped and reinstalled.