unify build and test scripts
Prior to this commit, dcos-metrics relied on several Makefiles and shell
scripts to build and test the components of this project. Beginning with
this commit, we attempt to unify these build scripts into a DRY
methodology.

This commit begins with the collector/ component, but these scripts will
be expanded to cover the Mesos module and consumers at some point as
well.