modify build script to test without errors
Prior to this commit, there were a few small issues with the build
script that weren't exposed until we resolved some other issues, like
fixing linting problems and so on.

This commit makes a few modifications that avoid passing variables
around as much, and builds the project (with coverage reports) without
returning errors.