This commit breaks out the kafka producer into its own package. In doing this we had to break out a few other features, notably: 	Events 		- Events package supports statsd event building
	Utils
		- Right now this is just the HTTP util stuff

This commit does not include the updated configuration, but does technically "build". In order to breakout the producer and events packages we needed to build out more robust configuration, this includes adding a statsd config and kafka config to the top level dcos-metrics-collector.go config struct. That will be the next, upcoming commit to this one.