remove statsd metrics emission for collector
Prior to this commit, various parts of the codebase could emit metrics
about internal workings. This included things like

  * TCPSessionOpened
  * TCPSessionClosed
  * AvroBytesIn
  * KafkaSessionOpened

and so on. However, this kind of metrics emission from the project
itself is outside of its scope and design and could be better suited as
log messages instead.

In this commit, I remove the StatsD metrics emission from Main and
Collector.