refactor http client into a more reusable solution
Prior to this commit, the HTTP client was broken up into several
different methods and wasn't tested at all. It also required that the
caller handle some errors itself, while the client handled others.

In this commit, I've refactored the HTTP client to be composable and
reusable. In addition to handling all the needs of this project, it also
allows callers to pass in a variable / struct that it will return data
to. This should be very helpful if the caller already has a struct with
JSON tags that the resulting JSON can map to.