finish developing generic MetricsMessage struct
Prior to this commit, some of the responses present in the HTTP producer
were actually good to have as part of the interface / struct of incoming
metrics, fields that all producers will need in order to produce
valuable output.

This commit moves those structs from the HTTP producer into the generic
MetricsMessage struct, ensuring a consistent experience for all producer
developers.