create janitor to remove stale objects from store
Janitor runs in its own goroutine and removes stale objects from the
in-memory store based on their "last updated" timestamp.

As objects are added to the in-memory store (via dcos-go/store), they
are created by their MetricsMessage name. If the metrics continue to be
seen, they'll be added to the store, overwriting their old value, thus
"updating" their timestamp.

If the metric has disappeared (the container existed, for example), that
object won't be overwritten in the store, and its timestamp won't be
"updated".

The Janitor looks at all objects every 60 seconds (hard-coded for now)
and, if the object is older than (pollingPeriod * 2) seconds, it will be
Delete() ed.