wait before attempting to query the http producer
When the HTTP producer is started in TestHTTPProducer_Agent,
it runs in its own goroutine. We're seeing some flakiness in CI (where
things might run a bit slower) where the query to the HTTP producer
fails with "connection refused." This doesn't happen all the time,
which leads me to believe it's a timing issue.

This commit adds a 1 second time.Sleep() immediately after the call to
pi.Run(), which I *hope* will resolve this. If not, I'll revert this
commit.