refactor framework collector to use a struct
Prior to this commit, RunFrameworkTCPListener() was a bare function in
collectors/framework/framework.go, with multiple configuration options
defined as top-level variables in that file. This had the unintended
side effect of not being able to set these configuration options
elsewhere, such as mocking them or modifying them for testing.

This commit refactors the framework collector, specifically
RunFrameworkTCPListener() to be a method on the framework.Collector
struct. It also adds a new function, New(), which returns a new
Collector{} struct with the default configuration. This patch will allow
us to mock this configuration, perhaps most importantly
listenEndpointFlag, during testing so that we can use a random,
ephemeral port on our test target.