fetch leader /state endpoint for fw principal
Prior to this commit, we only fetched state from the local agent, but
only the leader's /state endpoint can provide the framework principal.

This commit adds a method to query the master's /state endpoint for the
framework principal. It also adds a mocked cluster state object to
agent_test.go, and ensures that FrameworkPrincipal is set to the correct
value.