Created the Librato plugin.
Two additional flags are needed to run the plugin:

* librato-email
* librato-token

The plugin connectorFunc receives a bundle of MetricMessages
and then uses the postRequest type to send them to Librato.

Note that the connectorFun does not ever return nil.  This is
because it causes the plugin to terminate in the plugin
framework's current state.