Add --config flag all plugins
The --config flag is optional, and is a path to a yaml file with further
configuration. This rather complicated-seeming arrangement is useful
because it allows the plugins to share a config file with the
metrics agent proper, rather than passing explicit configuration items
like the CA certificate to the plugin directly.