Use configurable IAM in preference to JWT token in plugins (#99)
* Remove the --auth-token parameter

The --auth-token parameter allowed the user to generate a JWT and start
a plugin with the specified parameter. In systems which used such
tokens, the plugin would crash as soon as the JWT expired.

* Remove redundant MetricsAuthToken helper

* Add --config flag all plugins

The --config flag is optional, and is a path to a yaml file with further
configuration. This rather complicated-seeming arrangement is useful
because it allows the plugins to share a config file with the
metrics agent proper, rather than passing explicit configuration items
like the CA certificate to the plugin directly.

* Load yaml configuration into plugin

This commit extends the plugin struct with IAMConfigPath and
CACertificatePath (this is modeled after the Config struct in
dcos-metrics proper).

These two options are read in by the yaml reader from the path specified
by the Config flag.

* Don't attempt to read config file if not specified

* Hold an authenticated client on the plugin

calling createClient() will create a client with any configured IAM
options which the plugin can use in requests to the metrics API.

* Load in config and create the plugin on first run

This has to be done in Metrics() rather than StartPlugin() so that the
CLI options offered to the plugin are available.

* Use authenticated client in setEndpoints

* Use authenticated client in makeMetricsRequest

* Remove superfluous string default values

* Rename MetricsProto to MetricsScheme

* Avoid double-reference

* Update MetricsPort to be int, not string