Replace labels with tags
This removes the Label dimension from the MetricsMessage and applies all
previously held as a label as a tag on the datapoint instead. This
allows multiple messages with labels to be combined, since they are
simply datapoint tags now.

This is useful e.g. in kafka, where the
'kafka.network.RequestMetrics.LocalTimeMs.p999' metric can appear twenty
times with different tags.

Resolves DCOS_OSS-741