Accept sequential statsd input on app metrics channel (#100)
* Add app metrics to the store by name

* Retrieve named app metrics from the store

* Revert naive single-datapoint assumption

Previously, the store name of the message was qualified with the name of
the first datapoint. This worked well iff app metrics were all
dispatched individually. If app metrics were ever batched, as happens
even in kafka, the store would have held multiple datapoints in a single
message.

* Write all objects to store with a qualified name

This change updates the way that all objects are sent to the store,
rather than just the app metrics.

Where before (batched) node metrics were all stored in a single
'dcos.metrics.node' object, now each is stored individually using its
old name as a prefix and its datapoint name as a qualifier.

* Ensure one datapoint is written to store in test

The store now holds datapoint-qualified messages; if a message arrives
without datapoints it is discarded. We therefore add a datapoint to the
message in the test.

* Qualify metrics names with tags

By adding all tags into the metrics store identifier, we ensure metrics
with similar names (e.g. network.out.dropped) but with differing tags
(e.g. interface: lo / interface: docker0) are both retained, instead of
allowing one to overwrite the other and losing data.

* Replace labels with tags

This removes the Label dimension from the MetricsMessage and applies all
previously held as a label as a tag on the datapoint instead. This
allows multiple messages with labels to be combined, since they are
simply datapoint tags now.

This is useful e.g. in kafka, where the
'kafka.network.RequestMetrics.LocalTimeMs.p999' metric can appear twenty
times with different tags.

Resolves DCOS_OSS-741

* Introduce helper to combine metrics messages

Now that all metrics are held individually in the store, we need to be
able to retrieve them and combine them into a single message when they
are served.

* Use helper to combine app metrics

* Remove naive trimming from node metrics handler

Previously, the node metrics handler got all metrics associated with the
node ID and served the first. This was never great, and is now invalid.

* Use helper to combine node metrics

* Retrieve all container metrics in container handler

* Use helper to combine container metrics

* Retrieve all app metrics in app-metric handler

* Check for missing data, not error, in app-metric handler

* Use helper to combine named app metrics

* Dedupe container ids in /containers handler

* Escape inputs in regular expressions

* Move repeated string joins to helper function

* Improve example metric names

* Make it harder for tags to collide with metrics

We adopt the dogstatsd tag format to help avoid collions between tag
names and metric names. It's not perfect, since metric names might
contain hashes and colons, but it's better.

* Use switch, not if, for type assertion

* Use switch, not if, for container type assertion

* Throw 500, not 400, on store error

* Fetch metrics without wildcard

Previously we fetch metrics with the container id, a wildcard, and the
metrics id. There was no need for this wildcard; the metrics id follows
the container id directly and is followed by tags, not preceded by tags.

* Return 204, not 400, when no node metrics are found

* Clarify names in for loop

* Remove redundant nil check

* Prioritise tags over labels

* Warn on label/tag collision

* Assert that labels are trimmed from dimensions

* Add test for sequential stats input

* Use helper to join metrics names in handlers.go