Add task ID to /container endpoint (#114)
* Expand agent state model to process task info

Agent state is a huge JSON model (ref
http://mesos.apache.org/documentation/latest/endpoints/slave/state/).

We only model the parts of it we care about, which up until now did not
include tasks. This commit models the tasks object and the mapping
leading to the container ID embedded in each task.

* Replace executorLabels with stateLabel

`executorLabels` is an unnecesarily specific name for a struct with is a
simple key/value pair. In order to re-use it in task, the name is
broadened to `stateLabel`.

* Add valuable context comment to state model

* Add method to retrieve executor by ID

* Add method to retrieve task by container ID

* Test that no error is thrown when no task status is present

* Separate present and missing container metrics in tests

* Ensure test data executor IDs match container

These have always been out of synch with one another but it didn't
matter before; now it does.

* Remove outdated comment

* Add container status to agent state test fixture

* Test that task ID and name are returned as dimensions

* Return task ID and name as dimensions

* Rename stateLabel to keyValue

* Fix case in comment

* Replace struct, ok util return types with *struct

Instead of returning an empty struct from the utility methods in
agent.go, we now return a pointer to a struct. Instead of an !ok
response, failure is now indicated by a nil pointer.