Draw container IDs in /containers from both container and app metrics (#116)
* Update signature of test setup method

This will allow the same method to be used without sending all the test
data.

* Populate /containers from app and container metrics

This commit adds a test which supplies node and app metrics, but no
container metrics. This fails without the accompanying fix, which draws
app and container metrics from the store and extracts the set of unique
container IDs from them.

* Restore accidentally-removed build flag

I would normally squash this fix in, but a downstream DC/OS PR is
depending on a specific commit hash, so I'm not going to do so. It will
get squashed when we merge.

* Rename allTestData to allTestMessages

* Improve readability of containers test case