Add BeforeFunc and AfterFunc to plugins
This has the effect of exposing app.Before and app.After (see the
urfave/cli documentation for detail). These are essential for a plugin
which performs actions which are not a direct result of being passed
metrics.

In the case of the Prometheus plugin, it will allow us to start and stop
an http server process.