Refactor Metrics method for readability
My apologies for the meatiness of this commit, but there wasn't a good
way to break this bit down.

Instead of calling the rather curiously named setEndpoints and then
running a request for each endpoint, we now approach the problem as
three distinct problems:
1. Get node metrics
2. Get a list of every container ID
3. Retrieve metrics for each container

We have methods for each of these operations, which Metrics now invokes
in sequence.