(1.10) Provide same set of label dimensions for metrics with the same name (#169)
* Fix golint URL in Dockerfile

* Rework Collect to stop prometheus from dropping metrics with same FQname

* Fix not sanitizing all the label names

* Fix errors that occur while testing with go 1.11 (#166)

* Build with Go 1.8

This removes the use of type aliases, which were introduced in Go 1.9.

* Use map as set instead of slice

* Fix goimports errors

* Parse Prometheus response in test

This improves the test by parsing the Prometheus response into a data
structure, instead of searching the response for substrings.

* Test Prometheus output completely with multiple cases

* Sanitize metric and label names when creating Prometheus objects

* Factor grouping datapoints out of Collect()

* Replace tag with label

* Factor getting Desc labels out of Collect()

* Simplify collecting labels

* Improve variable names

* Improve readability

* Move datapointLabels struct to top of file

* Move helper functions to bottom of file

* Don't provide constLabels to NewDesc()

Constant labels are supposed to be constant among all Descs with the
same name, but we're inferring constLabels from the limited set of
metrics we're processing at the moment, which may be inaccurate.
Instead we treat all labels as variableLabels.

* Fix goimports error